+++ firebase/functions/src/index.ts
@@ -5,3 +5,4 @@ initializeApp();
 export { createStripeCheckoutSession } from './stripe-checkout';
 export { stripeWebhook } from './stripe-webhook';
 export { getStripeSession } from './stripe-session';
+export { onUserCreate } from './on-user-create';